# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HDMI Multiviewer Proxy",
    "description": "REST API for controlling UHD-401MV multiviewer via RS-232",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/output/resolution": {
      "post": {
        "tags": [
          "Output"
        ],
        "summary": "Set Resolution",
        "description": "Set output resolution.\n\nResolution values:\n- 1: 4096x2160p60\n- 2: 4096x2160p50\n- 3: 3840x2160p60\n- 4: 3840x2160p50\n- 5: 3840x2160p30\n- 6: 3840x2160p25\n- 7: 1920x1200p60RB\n- 8: 1920x1080p60\n- 9: 1920x1080p50\n- 10: 1360x768p60\n- 11: 1280x800p60\n- 12: 1280x720p60\n- 13: 1280x720p50\n- 14: 1024x768p60",
        "operationId": "set_resolution_api_output_resolution_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetResolutionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutputResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/output/hdcp": {
      "post": {
        "tags": [
          "Output"
        ],
        "summary": "Set Hdcp",
        "description": "Set HDCP mode.\n\n- 1: HDCP 1.4\n- 2: HDCP 2.2\n- 3: OFF",
        "operationId": "set_hdcp_api_output_hdcp_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetHDCPRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutputResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/output/video-mode": {
      "post": {
        "tags": [
          "Output"
        ],
        "summary": "Set Video Mode",
        "description": "Set video mode (ITC).\n\n- 1: Video mode\n- 2: PC mode",
        "operationId": "set_video_mode_api_output_video_mode_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetVideoModeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutputResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/output/vka": {
      "post": {
        "tags": [
          "Output"
        ],
        "summary": "Set Vka Pattern",
        "description": "Set video keep active pattern.\n\n- 1: Black screen\n- 2: Blue screen",
        "operationId": "set_vka_pattern_api_output_vka_post",
        "parameters": [
          {
            "name": "pattern",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Pattern"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutputResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OutputResponse": {
        "properties": {
          "resolution": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution"
          },
          "hdcp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hdcp"
          },
          "video_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Video Mode"
          },
          "vka_pattern": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vka Pattern"
          }
        },
        "type": "object",
        "title": "OutputResponse",
        "description": "Output settings response."
      },
...
      "SetHDCPRequest": {
        "properties": {
          "hdcp": {
            "type": "integer",
            "maximum": 3.0,
            "minimum": 1.0,
            "title": "Hdcp",
            "description": "1=HDCP 1.4, 2=HDCP 2.2, 3=OFF"
          }
        },
        "type": "object",
        "required": [
          "hdcp"
        ],
        "title": "SetHDCPRequest",
        "description": "Set HDCP request."
      },
...
      "SetResolutionRequest": {
        "properties": {
          "resolution": {
            "type": "integer",
            "maximum": 14.0,
            "minimum": 1.0,
            "title": "Resolution"
          }
        },
        "type": "object",
        "required": [
          "resolution"
        ],
        "title": "SetResolutionRequest",
        "description": "Set resolution request."
      },
...
      "SetVideoModeRequest": {
        "properties": {
          "mode": {
            "type": "integer",
            "maximum": 2.0,
            "minimum": 1.0,
            "title": "Mode",
            "description": "1=video, 2=pc"
          }
        },
        "type": "object",
        "required": [
          "mode"
        ],
        "title": "SetVideoModeRequest",
        "description": "Set video mode request."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}